requirements.txt
gspread==6.0.0
oauth2client==4.1.3
google-auth==2.27.0
google-generativeai==0.3.2
PyPDF2==3.0.1
requests==2.31.0
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
apscheduler==3.10.4
sendgrid==6.11.0